requirements.txt
akshare==1.0.52
backtrader==1.9.76.123
backtrader_plotting==2.0.0
dask==2.20.0
deap==1.3.1
dtshare==1.0.3
easytrader==0.23.0
interval==1.0.0
jqdatasdk==1.8.9
matplotlib==3.2.2
numpy==1.18.5
peewee==3.14.4
Pillow==9.1.0
plotly==5.1.0
psutil==5.7.0
PyAutoGUI==0.9.53
PyQt5==5.15.6
pyqtgraph==0.12.1
PySnooper==0.5.0
pytz==2019.3
pywin32==227
pywinauto==0.6.6
pyxll==5.4.1
pyzmq==22.3.0
QDarkStyle==2.8.1
quickfix==1.15.1
requests==2.26.0
rqdatac==2.9.31
schedule==1.1.0
scipy==1.5.0
TA_Lib==0.4.19
tigeropen==1.2.1
tqsdk==3.2.6
trading_calendars==2.1.1
tushare==1.2.62
twilio==6.62.1
tzlocal==2.1
vnpy_ctp==6.5.1.10
vnpy_ctptest==6.5.1.4
win10toast==0.9
win32gui==221.6
